+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * useIntegrationConfigs hook usage from stripes-acq-components. Refs UIORGS-281.
 * useOrganization hook usage from stripes-acq-components. Refs UIORGS-282.
 * Validation for "Account number" must be unique for organization. Refs UIORGS-292.
+* Adding tenant's timezone in Organization integration form.Refs UIORGS-300.
 
 ## [3.0.1](https://github.com/folio-org/ui-organizations/tree/v3.0.1) (2021-11-02)
 [Full Changelog](https://github.com/folio-org/ui-organizations/compare/v3.0.0...v3.0.1)